+++ uTransnet-Calc/Assets/uTrans/Scripts/Data/LinkDTO.cs
@@ -5,10 +5,10 @@ namespace uTrans.Data
     public class LinkDTO : BaseDTO
     {
         [PrimaryKey, AutoIncrement]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Indexed]
-        public long ProjectId  { get; set; }
+        public int ProjectId  { get; set; }
